fix: R37 第二轮（领导制）R3 轮——Windows 构建中断 + 配置根写提权原语 + 中继/日志韧性等
FIND-HEAD-R3（15 名子代理，冻结树 4d470f5）+ FIX-HEAD-R3（9 个工作组）：
R3 新 effective 30 条（H1/M7/L22）+ R2/R1 残留 31 条 → 本轮完成 22 + 23，
留待下一轮 9 条（含需解冻才能改的 PROTOCOL.md 文档偏差 L8/L10）；引入 0 新缺陷。

HIGH:
- auth.c 的 fprintf 用 %zu，mingw msvcrt printf 不支持 → CI win-cross（IWAN_WERROR=ON）
  rc=1，Windows 交付物全断供（由上一轮 M7 修复引入，blame 确认）。修：%llu + 显式转换，
  并全树清扫 %z 用法。**win 交叉构建 rc=0 零诊断**。
- 同根治理：为自家 logger 补 format(printf) 属性（此前 258 处日志格式零编译期检查，
  正是 H1 得以潜伏的机理）+ 阴性对照。

MEDIUM（要点）:
- oidc_config 根写守卫：符号链接/分量绕过仍可让 root 在任意位置新建任意名字目录并把
  属主交给调用者（提权原语）。修：分量 lstat + 目标类型检查 + realpath 仅用于 root 判定
  （fail-closed），动态 before/after。
- relay：握手预算对"发 1 字节后静默"不生效（256 半开连接永久钉死中继 → 0/256）；
  accept poll 瞬时错误杀线程；方向线程死亡后新注册连接无人 retire（改成 cleanup 唯一
  退役点 + 同锁死亡标记，512 次死亡泄漏 1024 fd → 0）；上限处谎报 SOCKS5 成功改拒绝。
- socks：accept 非 EAGAIN 活锁 + 日志洪泛（98-102% CPU）；EMSGSIZE/EINVAL 不再误判为
  死隧道触发重连风暴；EINTR 分类。
- exec_sanitize 补清 OPENSSL_CONF/MODULES/ENGINES（sudo 重执行前的 provider 加载面）。
- https/port/util/tun：指针溢出 UB、EINTR 无界重试吃掉超时预算、空值 header 泄漏、
  ip_run 对失败命令 fail-open 返回 true、set_nonblock 吞错、EINTR 绕过写预算。
- 可观测性：日志转义补 C1 控制字节（可伪造日志行）；错误信息带字节数。
- 文档/构建：README 与 --help 不一致、CMakeLists 普通 set() 遮蔽 cache 变量
  （-DCMAKE_RUNTIME_OUTPUT_DIRECTORY 实测无效）、Windows 严格门禁平台分档前置登记。

验收（最终树）：build / build-dbg / build-wstrict 三构建 rc=0 零诊断；
Windows x86_64 交叉构建（-DIWAN_WERROR=ON）rc=0 零警告；
免 root 夹具全绿（lwip_data、lwip_fairness 8×1MiB min==max、socks_handshake --token
21/21 与 --fail-max 1、socks_ipv6_e2e PASS）。无 sudo ⇒ 真 TUN/integration.sh 未跑。

+++ src/common/pump_win_single.c
@@ -53,8 +53,9 @@ void pump_win_single_pkt(void *ud, uint8_t *pkt, size_t len, bool last)
     if (len == 0 || len > PUMP_SLOT) {
         if (len > PUMP_SLOT)
             atomic_fetch_add(&g_prof_tun_rdrop, 1);
-        log_debug("pump: drop packet, len %zu out of [1, %d]", len,
-                  PUMP_SLOT);
+        /* %llu not %zu: msvcrt printf (Windows) lacks %zu */
+        log_debug("pump: drop packet, len %llu out of [1, %d]",
+                  (unsigned long long)len, PUMP_SLOT);
         return;
     }
     if (g_stop)
